refactor. 페이지 내부 컴포넌트 추상화 레벨 통일 & contextAPI 추가
